refactor: created view, infrastructure directories
