+++ app/src/main/AndroidManifest.xml
@@ -20,7 +20,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.Thermometer">
-        <activity android:name=".activity.MainActivity">
+        <activity android:name="jp.aoyama.mki.thermometer.view.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
