--- app/src/main/java/jp/aoyama/mki/thermometer/activity/MainActivity.kt
@@ -1,44 +1,0 @@
-package jp.aoyama.mki.thermometer.activity
-
-import android.content.Intent
-import android.os.Bundle
-import android.view.Menu
-import android.view.MenuItem
-import androidx.appcompat.app.AppCompatActivity
-import jp.aoyama.mki.thermometer.R
-import jp.aoyama.mki.thermometer.util.CSVFileManager
-
-
-class MainActivity : AppCompatActivity() {
-    private val mCsvFileManager = CSVFileManager()
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-
-    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
-        menuInflater.inflate(R.menu.share_menu, menu)
-        return super.onCreateOptionsMenu(menu)
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        when(item.itemId){
-            R.id.share -> shareFile()
-        }
-        return super.onOptionsItemSelected(item)
-
-
-    }
-
-    private fun shareFile() {
-        val file = mCsvFileManager.getFileUrl(this)
-        val shareIntent = Intent().apply {
-            action = Intent.ACTION_SEND
-            putExtra(Intent.EXTRA_STREAM, file)
-            addFlags(Intent.FLAG_GRANT_WRITE_URI_PERMISSION or Intent.FLAG_GRANT_READ_URI_PERMISSION)
-            type = "text/plain"
-        }
-        startActivity(Intent.createChooser(shareIntent, getString(R.string.share_file_description)))
-    }
-}
