+++ app/src/main/java/jp/aoyama/mki/thermometer/domain/models/User.kt
@@ -1,0 +1,9 @@
+package jp.aoyama.mki.thermometer.domain.models
+
+import java.util.*
+
+data class User(
+    val id: String = UUID.randomUUID().toString(),
+    val name: String,
+    val bluetoothData: BluetoothData?
+)
