--- app/src/main/java/jp/aoyama/mki/thermometer/models/BluetoothData.kt
@@ -1,13 +1,0 @@
-package jp.aoyama.mki.thermometer.models
-
-data class BluetoothData(
-    val name: String?,
-    val address: String,
-) {
-    companion object {
-        fun create(name: String?, address: String?): BluetoothData? {
-            if (address == null) return null
-            return BluetoothData(name = name, address = address)
-        }
-    }
-}
