--- app/src/main/java/jp/aoyama/mki/thermometer/models/User.kt
@@ -1,9 +1,0 @@
-package jp.aoyama.mki.thermometer.models
-
-import java.util.*
-
-data class User(
-    val id: String = UUID.randomUUID().toString(),
-    val name: String,
-    val bluetoothData: BluetoothData?
-)
