+++ app/src/main/java/jp/aoyama/mki/thermometer/view/bluetooth/list/BluetoothDiffUtil.kt
@@ -1,0 +1,14 @@
+package jp.aoyama.mki.thermometer.view.bluetooth.list
+
+import android.bluetooth.BluetoothDevice
+import androidx.recyclerview.widget.DiffUtil
+
+class BluetoothDiffUtil : DiffUtil.ItemCallback<BluetoothDevice>() {
+    override fun areItemsTheSame(oldItem: BluetoothDevice, newItem: BluetoothDevice): Boolean {
+        return oldItem.address == newItem.address
+    }
+
+    override fun areContentsTheSame(oldItem: BluetoothDevice, newItem: BluetoothDevice): Boolean {
+        return oldItem.address == newItem.address
+    }
+}
