+++ app/src/main/java/jp/aoyama/mki/thermometer/view/models/UserEntity.kt
@@ -1,0 +1,16 @@
+package jp.aoyama.mki.thermometer.view.models
+
+import jp.aoyama.mki.thermometer.domain.models.BluetoothData
+import jp.aoyama.mki.thermometer.domain.models.User
+import java.util.*
+
+data class UserEntity(
+    val id: String = UUID.randomUUID().toString(),
+    val name: String,
+    val bluetoothData: BluetoothData?,
+    val rssi: Int?,
+    val lastFoundAt: Calendar?,
+) {
+    constructor(user: User, rssi: Int?, lastFoundAt: Calendar?) :
+            this(user.id, user.name, user.bluetoothData, rssi, lastFoundAt)
+}
