+++ app/src/main/res/layout/select_name_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragment.user.list.SelectNameFragment">
+    tools:context=".view.user.fragments.SelectNameFragment">
 
 
     <LinearLayout
